+++ Makefile
@@ -1,4 +1,4 @@
-VPATH = common frontend test_cases arch/arm32
+VPATH = common frontend test_cases arch/arm32 backends/riscos
 
 COMMON =\
 	stream.c \
